fix authorisation

+++ src/modules/users/services/UrlService.js
@@ -1,34 +1,47 @@
-import UrlRepository from '../repositories/UrlRepository'
+import UrlRepository from '../repositories/UrlRepository';
 import NotFound from '../../../classes/errors/4xx/notFound'
 
 class UrlService {
-
-    async list() {
-        return await UrlRepository.readAll()
-    }
 
     async create(url) {
         return await UrlRepository.create(url)
     }
 
-    async findById(id) {
+    async readAll() {
+        return await UrlRepository.readAll()
+    }
+
+    async readById(id) {
+
         let url = await UrlRepository.readById(id)
+
         if (!url) {
-            throw new NotFound(`${objectName} not found`)
+            throw new NotFound('Url not found')
         }
+
         return url
     }
 
     async update(id, url) {
-        let nUrl = UrlRepository.readById(id)
-        if(!nUrl) {
-            throw new NotFound(`${objectName} not found`)
+
+        let oldUrl = await UrlRepository.readById(id)
+        
+        if (!oldUrl) {
+            throw new NotFound('Url not found')
         }
+
         return await UrlRepository.update(id, url)
     }
 
     async destroy(id) {
-        await UrlRepository.destroy(id)
+
+        let url = await UrlRepository.readById(id)
+        
+        if (!url) {
+            throw new NotFound('Url not found')
+        }
+        
+        return await UrlRepository.destroy(id)
     }
 }
